// generated/src/app/api-models/cap-cputl-rpt-tmp-wrk.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapCputlRptTmpWrk {
  provId: number;
  vendId: number;
  seqCapPoolId: number;
  seqClaimId: number;
  claimFileType: string;
  seqGroupId: number;
  applyTo: string;
  capEntityCode: string;
  capTransAmt: number;
  pmpmWithholdAmt: number;
  seqCapCputlRptTmpWrk: number;
  seqMembId: number;
  capMonth: Date;
  adjustmentFlag: string;
}